data-guardian/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
pathlib>=1.0.1

